+++ demo/src/app/app.module.ts
@@ -12,10 +12,13 @@ import { routing } from './app.routing';
 import { AppComponent } from './app.component';
 import { CaseUIToolkitModule, DraftService, AlertService, HttpService, AuthService, CasesService,
   HttpErrorService, AbstractAppConfig, CaseEditWizardGuard, RouterHelperService,
-  PlaceholderService, DocumentManagementService, PageValidationService} from '@hmcts/ccd-case-ui-toolkit';
+  PlaceholderService, DocumentManagementService, PageValidationService, SearchService,
+  RequestOptionsBuilder} from '@hmcts/ccd-case-ui-toolkit';
 import { ScrollToService } from '@nicky-lenaers/ngx-scroll-to';
 import { CaseProgressConsumerComponent } from './case-progress-consumer.component';
 import { CoreComponent } from './core.component';
+import { CaseViewConsumerComponent } from './case-view-consumer.component';
+import { SearchFiltersConsumerComponent } from './search-filters-consumer.component';
 
 @NgModule({
   imports: [
@@ -35,10 +38,14 @@ import { CoreComponent } from './core.component';
     CoreComponent,
     CaseCreateConsumerComponent,
     CaseProgressConsumerComponent,
+    CaseViewConsumerComponent,
+    SearchFiltersConsumerComponent
   ],
   providers: [
     CasesService,
     AuthService,
+    SearchService,
+    RequestOptionsBuilder,
     HttpService,
     HttpErrorService,
     AlertService,