+++ src/shared/components/case-viewer/case-event-trigger/case-event-trigger.component.ts
@@ -3,9 +3,10 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Subscription } from 'rxjs/Subscription';
 import { DisplayMode, CaseEventTrigger, CaseView, Activity, CaseEventData } from '../../../domain';
-import { CasesService } from '../../case-editor';
+import { CaseService, CasesService } from '../../case-editor';
 import { AlertService, ActivityPollingService, EventStatusService } from '../../../services';
 import { CaseReferencePipe } from '../../../pipes';
+import { NgZone } from '@angular/core';
 
 @Component({
   selector: 'ccd-case-event-trigger',
@@ -16,9 +17,12 @@ export class CaseEventTriggerComponent implements OnInit, OnDestroy {
   eventTrigger: CaseEventTrigger;
   caseDetails: CaseView;
   subscription: Subscription;
+  parentUrl: string;
 
   constructor(
+    private ngZone: NgZone,
     private casesService: CasesService,
+    private caseService: CaseService,
     private router: Router,
     private alertService: AlertService,
     private route: ActivatedRoute,
@@ -27,15 +31,30 @@ export class CaseEventTriggerComponent implements OnInit, OnDestroy {
   ) { }
 
   ngOnInit(): void {
-    this.caseDetails = this.route.snapshot.data.case;
+    if (this.route.snapshot.data.case) {
+      this.caseDetails = this.route.snapshot.data.case;
+    } else {
+        this.caseService.caseViewSource.asObservable().subscribe(caseDetails => {
+          this.caseDetails = caseDetails;
+        });
+    }
     this.eventTrigger = this.route.snapshot.data.eventTrigger;
-    this.subscription = this.postEditActivity().subscribe((_resolved) => {
-      // console.log('Posted EDIT activity and result is: ' + JSON.stringify(resolved));
+    if (this.activityPollingService.isEnabled) {
+      this.ngZone.runOutsideAngular( () => {
+        this.subscription = this.postEditActivity().subscribe((_resolved) => {
+          // console.log('Posted EDIT activity and result is: ' + JSON.stringify(_resolved));
+        });
+      });
+    }
+    this.route.parent.url.subscribe(path => {
+      this.parentUrl = `/${path.join('/')}`;
     });
   }
 
   ngOnDestroy() {
-    this.subscription.unsubscribe();
+    if (this.activityPollingService.isEnabled) {
+      this.subscription.unsubscribe();
+    }
   }
 
   postEditActivity(): Observable<Activity[]> {
@@ -55,7 +74,7 @@ export class CaseEventTriggerComponent implements OnInit, OnDestroy {
   submitted(event: any): void {
     let eventStatus: string = event['status'];
     this.router
-      .navigate(['case', this.caseDetails.case_type.jurisdiction.id, this.caseDetails.case_type.id, this.caseDetails.case_id])
+      .navigate([this.parentUrl])
       .then(() => {
         let caseReference = this.caseReferencePipe.transform(this.caseDetails.case_id.toString());
         if (EventStatusService.isIncomplete(eventStatus)) {
@@ -68,7 +87,10 @@ export class CaseEventTriggerComponent implements OnInit, OnDestroy {
   }
 
   cancel(): Promise<boolean> {
-    return this.router.navigate(['/case', this.caseDetails.case_type.jurisdiction.id, this.caseDetails.case_type.id,
-      this.caseDetails.case_id]);
+    return this.router.navigate([this.parentUrl]);
+  }
+
+  isDataLoaded(): boolean {
+    return this.eventTrigger && this.caseDetails ? true : false;
   }
 }