+++ src/shared/components/index.ts
@@ -7,3 +7,5 @@ export * from './error';
 export * from './event-trigger';
 export * from './markdown';
 export * from './palette';
+export * from './search-filters';
+export * from './create-case-filters';