+++ src/shared/domain/definition/case-field.model.ts
@@ -1,11 +1,13 @@
-import { Orderable } from '../order/orderable.model';
+import { Orderable } from '../order';
 import { FieldType } from './field-type.model';
 import { WizardPageField } from '../../components/case-editor/domain';
 import { Type } from 'class-transformer';
+import { AccessControlList } from './access-control-list.model';
 
 // @dynamic
 export class CaseField implements Orderable {
   id: string;
+  hidden?: boolean;
   label: string;
   order?: number;
 
@@ -19,6 +21,7 @@ export class CaseField implements Orderable {
   show_condition?: string;
   show_summary_change_option?: boolean;
   show_summary_content_option?: number;
+  acls?: AccessControlList[];
 
   @Type(() => WizardPageField)
   wizardProps?: WizardPageField;